fix(sched): memory corruption due to popping context that was never pushed
Task object of the next task was corrupted, because RESTORE_CONTEXT_FROM_IRQ
was used pop context from the stack of the to-be scheduled task. But on
the first time the task is run, there is no previous context on the stack.
This caused the sp to point past the allocated space, into the next task's
struct, and corrupting it. Also added padding to stacks to alleviate potential
future overflows.